Moving CSS into its own file.
This really has no major performance benefits. Ideally we would inject
an iFrame with the overlay's HTML and CSS in its own sandbox, but you
can't mess with stuff in an iFrame that's not on the same domain as its
parent. But moving the CSS into its own file keeps it from showing up
twice on the page (once in end.js, and once again when it's put into the
page via document.write), and makes our document.write for the overlay
far more digestible.