Tweaking Overlay UI
- At the expense of another div, the link display now retains its proper
proportions whilst scaling to the window width and remaining centered at
all times.  - eventListener's for the overlay are now added only once at
the injection of the overlay HTML, as opposed to every time
showURLinWindow() is invoked.  - setTimeouts on the showing and hiding
of the overlay were tweaked so that they don't clip the animation.  -
popover.html was deleted as it isn't used by the extension (it was a
remnant from experiments with Safari extension popovers).