Drastically improving fragment link generation.
Huge improvements to fragment link generation. We now check a far larger
amount of identifiers, and attempt to create the most concise and unique
(in that order) identifier for an element.

To the other plugin authors: This is basically entirely new logic, but
it should hopefully be easy to drop in to your code, if you'd like. If
you've got any questions, contact me and I'll do my best to help you out.
And if you've got any improvements or bug fixes, definitely let me know.
Thanks!

Other notes:
 - Rewrote singleElementWithID(). It's now singleElementWithSelector(),
and uses querySelectorAll() instead of messing around with the DOM
stupidly. It's also faster for the majority of its uses.

 - Added dictionaryToSelector(). It's used to turn a dictionary (or
object literal, in javascript parlance) into a selector string.

 - Modified the check for selector uniqueness per Lim Chee Aun's
for improved speed. The check is now called by firstElementWithSelector().