Fixing CSS being overwritten by a website's stylesheet.
This is a pretty terrible fix, but I've just made all styles !important.
Injected stylesheets in Safari extensions are treated as user
stylesheets, so this seems to be the best I can do with this method. It
shouldn't affect websites' CSS, unless they're using the same IDs for
their elements. I'll be coming up with a better solution.