Rewriting highlighting UI to mimic Safari's search highlighting
The Good News: The animations are entirely CSS3, so I don't have to do anymore
manual animation, or force the user to load jQuery every time they use a
fragment link. This also looks way nicer than the yellow fade I was using
earlier, and acts almost exactly like Safari and Lion's highlighting (save for
a few intentional changes), so non tech-savvy users won't be left confused. It
also cuts down on the amount of javascript I'm loading into every page, and
removes an option from the preferences.

The Bad News: There are all sorts of ways in which this breaks. I'd say for
the most part it works well, and then it works okay and terribly in order of
frequency. For example:

     Good: http://avc.lu/rolgO6

     Okay: http://breakingbad.wikia.com/wiki/Breaking_Bad_Wiki#css(p%3Espan)
(Here #WikiaHeader's z-index seems to mess with getBoundingClientRect().
Setting the z-index to 0 or auto fixes the problem, but I struggle to find a
good way to handle that programmatically.)

     Terrible: http://shauninman.com/pilation/#css(img%23shaun) (Here the
issue is that CSSFrag interrupts jQuery's animations, which then stop
completely. I can't figure out a good way to detect when jQuery *isn't*
animating, though.)

For now I'm going to try to get the code cleanup and URL shortener support out
the door, so that all the features of CSSFrag will be operational. Then it'll
be back to trying to fix these bugs, and the others I keep finding.

(We're actually not replicating Safari's search highlighting, but Lion's
system-wide dictionary highlighting. The simple reason is because the latter
doesn't require an image in order to replicate it. Safari's search
highlighting has a gradient background and border, with a 3px radius. The only
way to replicate a gradient background *and* border with just one element is
to use -webkit-border-image and an image. Lion has a gradient background and a
solid border, with a 5px radius. Since the border is one solid color, we can
replicate the entire look with one element and no images.)