Fixing incorrect highlighting due to imgs with height and width of 0 bug
When a height and width property aren't explicitly set for an img,
getComputedStyle() reports 0 instead of the height and width of the
image, and getBoundingClientRect() is also therefore thrown off. In this
case, we now make a copy of the image, and copy its dimensions onto the
img element, fixing the issue. We walk up the entire tree from the
element we wish to highlight, checking all siblings for img elements,
and repairing their height and width if necessary.