Fixing scrollFocusAndHighlight() halting jQuery animations
Holy mother of god, is this ever a hack. Strap in boys and girls, here's how we
run this circus:

    1) The user loads a CSSFrag URL.
    2) handleLoadAndHashChange() is called via the window.load listener, with
just the event argument; isSafe is undefined.
    3) handleLoadAndHashChange() calls scrollFocusAndHighlight(), with the
arguments selector, and isFragHash; isSafe is still undefined.
    4) scrollFocusAndHighlight(), seeing that isSafe is undefined, injects some
javascript into the head of the document, which runs a function on a loop,
which basically uses a counter to try to guess when animations have finished
running (if any are running at all).

(Sidenote here: Why is there now way in jQuery to check if any animations are
running?)

    5) Once the injected javascript thinks that no animations are still
running, it fires off a custom event, "animationFinished".
    6) handleAnimationFinished() listens for this event, and when it is fired,
handleAnimationFinished() calls handleLoadAndHashChange() with what is a
useless event (because part of the Safari extensions API is apparently silently
stripping out any custom content in event objects. I found this out the time
consuming way), but with the isSafe set to true.
    7) handleLoadAndHashChange() runs through its entire set of instructions
again, but this time calls scrollFocusAndHighlight() with isSafe set to true.
    8) scrollFocusAndHighlight(), with isSafe set to true, bypasses the
animation check and runs its code as normal, scrolling to and highlighting the
element on the page.

Practically none of this insanity would be required if Safari extensions'
injected javascript (like the entirety of end.js) resided in the same scope as
all the other scripts loaded on the page. They of course don't, for security.

A little less insanity would be required if Safari did not silently strip the
custom attributes attached to an Event object when they're being sent to
extensions. They of course do, for security.

The last five hours of my life could have been spent pondering on the meaning
of life, writing a great American novel, or developing more features for
CSSFrag if Apple had documented that they silently strip custom attributes
attached to Event objects. They of course didn't, because they hate me. Or I'm
a terrible reader and this is all my fault.

Either way, I fixed the bug. So now
http://shauninman.com/pilation/#css(img%23shaun) works perfectly. Next up,
checking to see if other javascript frameworks have this same issue.