Refactoring plugin for sanity, and free bug fixes
Note that this commit is non atomic, and this developer is very sorry.

animationChecks.js isn't loaded by the extension. It's just so I can
work on those checks with proper syntax highlighting and no string
wrapping cruft. The checks are compressed and stored in end.js, since
there is no way for it to access animationChecks.js.