Adding link shortening support
Note that all instances of "URL" are now "link", for consistency with the UI.

The shortening support works, it's just somewhat lacking. Safari doesn't like it
when you make a request for just text data, so if the service doesn't return
JSON, I currently can't use it. Additionally, post requests are subject to all
the standard rules, so goo.gl support seems to be near impossible.

As for the API key for bit.ly, it's required in lieu of OAuth or XAuth support.
OAuth is probably impossible due to extension sandboxing, but XAuth is possible,
though since bit.ly seems to be the only service with support for it, it's not
exactly low hanging fruit.

Regardless, I'll be looking at improving the flow for and the number of link
shortening services in the future.