Fixing highlighting not picking up italics, bolds, etc.
The issue was because I only copied the full style of the selected
element, and not the styles of its children. I've solved the problem
with cloneStyles(), which copies the full style of an element and its
children to a cloned element. It does so recursively, however, but given
the nature of how I expect fragment links are used, there shouldn't be
that many elements to traverse.