Fixing fragment link referencing CSSFrag classes
A downside to cloning all elements in a selector is that the inline
styles override pointer-events: none. So now we force that to none
again.  Additionally, I had commented out the code that removed the
#CSSFragHighlightBackground element, and forgot to uncomment it. Which
is how I found this bug in the first place. So now I've fixed that, too.