Redrawing icons
I lost the original PSDs due to a slip of the rm, so I had to redraw the icon to produce new sizes for Chrome (Firefox and Opera developers, contact me if you need sizes I haven't provided). I've updated the extension's icons with the redrawn version, so everything should be in sync. The PNGs have already been optimized, so they should be as small as they can get.

A note on the file names: The first number is the size of the image itself, whereas the second number in parenthesis is the size of the icon within the image. If there's no second number, the size of the icon is the same as the image.