Fixing prefer standard hashes bug and link shortening of standard hashes bug
Previously, the prefer standard hashes setting was ignored (more
specifically, the check for a valid target id was missing "#", and so
always failed), and if a standard hash was generated, link shortening
would never take place (it was inside the else conditional instead of
outside). Both these bugs have been fixed, and the extension should now
do what it says it does.