Fixing competing user styles bug
Since CSSFrag inserted its stylesheet manually (in an effort to cut down on the amount of data loaded into every page), Safari didn't treat its stylesheet as a user style. This meant that user styles would take precedence over any bit of CSS CSSFrag attempted to insert, which mucks up highlighting in some cases (If you have Beautipedia, for instance, almost all highlights on wikipedia will be off by 13px).

Unfortunately, the only way to solve this is by having Safari insert the stylesheet into every page. This really isn't that terrible, though, since the CSS is only 3kb, which pales in comparison to the 15kb of javascript that CSSFrag has to dump into every page. And as all of CSSFrag's selectors are prefixed with CSSFrag, I would think that it shouldn't interfere with the styling of any pages

Note that this also removes the need for injectStyles(), which was previously responsible for injecting CSSFrag's stylesheet into the page, so it and all references to it have been removed.