Rewriting checkforAnimation()'s jQuery check to only check the selected element for animation
Previously, we'd check all elements on the page for animation, and only
highlight the selected element when we thought all animations on the
page were finished. But if a page had an element that was always
animating, we'd never highlight the selected element. So now we just
check the selected element for animation, as I'm pretty sure jQuery's
:animated pseudo selector applies to children of an animating element as
well. If the selected element is constantly animating, we'll still never
highlight it, but I assume this to be a rare use case.