Adding bug to the Readme
I seriously have no idea where this bug comes from. #CSSFragHighlightBackground, whose height and width are set to auto, will end up with larger dimensions (excluding the padding and border) than #CSSFragHighlight's, which shouldn't be possible. In the example in the read me, the highlighted element has no padding, border, or margin that could possibly account for the discrepancy. Web inspector's view of the two elements' box models show different height for each, but no margin, border, or padding for the highlighted element.

If anyone has can figure out the solution to this problem, I'll be greatly indebted to them. I can't fathom how I'm going to solve this one.