Fixing highlighting invisible elements
Specifically, I'm solving two problems. The first is highlighting elements with an opacity of 0. I just force it to 1, which isn't necessarily what the author intended, but doesn't confuse the user and lasts only 1.6 seconds anyway.

Secondly, I'm attempting to solve an issue with elements who are set to "display: none" when not being hovered over. To solve this, CSSFrag checks to see if the element is displaying, and if not fakes a mouseover event, and copies the element's hovered style. This is better than using getComputedStyle with a pseudo selector, as the latter case can't handle elements having their styles changed by anything other than the stylesheet.

Both these fixes mean that now this link works: http://shauninman.com/pilation/#css(img.logo)