Changing CSSFrag to Pinpoint
CSSFrag is now Pinpoint. Fragment links are now Pinpoints, and you create them instead of generating them. generateFragmentLink() is now createPinpoint(). There's a new icon, with a pin instead of a link symbol. I think the whole thing sounds better, and should be much friendlier to people who have no idea what CSS or fragment links are.

If there are no objections, I'd like this to be the nomenclature for this system, as I think a more approachable name that describes what these extensions/plugins *do* rather than how they do it is likely to see wider adoption (my mother near instantly understood what Pinpoint did. Explaining CSSFrag was…difficult).

The one change I'm rather on the fence about is changing #css() to #pinpoint(). On the side of the latter, people who don't have the extension or plugin for their browser can piece together a google search that would point them to where they can download it ("pinpoint link" versus "CSS link"). But on the side of the former, it respects the draft, and it makes a bit more semantic sense. If we were to switch, all the extension and plugin developers would need to update our code to reflect the change, and add a line or two for backwards compatibility. We can't have anything other than all of us making the change, otherwise we'll be introducing fragmentation, which I for one certainly wouldn't want.