Changing pinpoints (as in the links) to pins.
This'll get rid of sentences like "And since Pinpoint always creates the
most compatible pinpoint, in many cases your pinpoint will take anyone
to the part of the page you wanted to show them, even without
Pinpoint.", and honestly it was pretty stupid to call the links the
plural of the extension name.  Note: createPinpoint() is now
createPin(), for obvious reasons.  MILDLY IMPORTANT: Sometime later
tonight or tomorrow, I'll be simultaneously putting up a site for
pinpoint, renaming this repo to pinpoint, and updating users to the new
extension with the new nomenclature. These changes are purely
superficial, and the extension's workings and syntax will not be
modified, so it will still be compatible with the extensions/plugins
being developed for other browsers.