Tweaking highlighting to better match Lion
The scaling factor in Lion is about 1.425, not 1.25. It also starts at a
scale of 1 instead of 0. But this looks a little bit weird with large
elements, so we still start at a scale of 0, but fade the opacity from 0
to 1 by the time the element is at a scale of .7125, which looks better.
The time to scale down to a scale of 1 has been increased as well, again
because it looks weird on large elements otherwise.